oauth refactored and added to all routes
