+++ backend/api/document.py
@@ -7,8 +7,11 @@ from sqlalchemy.orm import Session
 
 from backend.db.sessions import get_db
 from backend.exceptions import NoDocumentsFoundException
+from backend.exceptions import NoValidPermissionsException
+from backend.models.sqlalchemy_models import User
 from backend.service.document_service import document_list
 from backend.service.document_service import save_document
+from backend.service.oauth import get_current_user
 
 # from fastapi.responses import JSONResponse
 # from typing import List
@@ -27,7 +30,9 @@ document_router = APIRouter(
 
 @document_router.post("/upload")
 async def upload_file(
-    file: UploadFile = File(...), db: Session = Depends(get_db)
+    current_user: User = Depends(get_current_user),
+    file: UploadFile = File(...),
+    db: Session = Depends(get_db),
 ):
     """
     This endpoint allows you to upload a single file.
@@ -35,7 +40,10 @@ async def upload_file(
     and content type as a JSON response.
     """
     try:
-        return save_document(file=file, db=db)
+        return save_document(current_user=current_user, file=file, db=db)
+
+    except NoValidPermissionsException as e:
+        raise HTTPException(status_code=403, detail=str(e))
     except Exception as e:
         raise HTTPException(
             status_code=500,
@@ -45,13 +53,18 @@ async def upload_file(
 
 
 @document_router.get("/")
-def list_documents(db: Session = Depends(get_db)):
+def list_documents(
+    current_user: User = Depends(get_current_user),
+    db: Session = Depends(get_db),
+):
     """List all the documents which exist in db"""
     try:
-        return document_list(db)
+        return document_list(current_user, db)
 
     except NoDocumentsFoundException as e:
         raise HTTPException(status_code=404, detail=str(e))
+    except NoValidPermissionsException as e:
+        raise HTTPException(status_code=403, detail=str(e))
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
